3d rendering - cube tex coords not accurate
Tex coords are only given for the front and back faces of the cube, not the side faces. This can be most easily fixed by just using 36 vertices without element buffers instead of 8 vertices. Will likely make a set of primitive meshes for these sorts of things where a model file isn't needed